requirement.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
coverage==7.8.0
cryptography==45.0.2
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
graphviz==0.20.3
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlalchemy_schemadisplay==2.0
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
google-cloud-storage==2.10.0
